+++ src/sentry/api/serializers/models/user.py
@@ -25,10 +25,10 @@ from sentry.models.organizationmapping import OrganizationMapping
 from sentry.models.organizationmembermapping import OrganizationMemberMapping
 from sentry.models.useremail import UserEmail
 from sentry.models.userpermission import UserPermission
-from sentry.models.userrole import UserRoleUser
 from sentry.organizations.services.organization import RpcOrganizationSummary
 from sentry.users.models.authenticator import Authenticator
 from sentry.users.models.user import User
+from sentry.users.models.userrole import UserRoleUser
 from sentry.users.services.user import RpcUser
 from sentry.utils.avatar import get_gravatar_url
