+++ src/sentry/backup/services/import_export/impl.py
@@ -48,9 +48,9 @@ from sentry.hybridcloud.models.outbox import outbox_context
 from sentry.models.importchunk import ControlImportChunk, RegionImportChunk
 from sentry.models.organizationmember import OrganizationMember
 from sentry.models.userpermission import UserPermission
-from sentry.models.userrole import UserRoleUser
 from sentry.silo.base import SiloMode
 from sentry.users.models.user import User
+from sentry.users.models.userrole import UserRoleUser
 
 logger = logging.getLogger(__name__)
