+++ src/sentry/testutils/factories.py
@@ -141,7 +141,6 @@ from sentry.models.team import Team
 from sentry.models.useremail import UserEmail
 from sentry.models.userpermission import UserPermission
 from sentry.models.userreport import UserReport
-from sentry.models.userrole import UserRole
 from sentry.organizations.services.organization import RpcOrganization, RpcUserOrganizationContext
 from sentry.sentry_apps.apps import SentryAppCreator
 from sentry.sentry_apps.installations import (
@@ -168,6 +167,7 @@ from sentry.uptime.models import (
     UptimeSubscription,
 )
 from sentry.users.models.user import User
+from sentry.users.models.userrole import UserRole
 from sentry.users.services.user import RpcUser
 from sentry.utils import loremipsum
 from sentry.utils.performance_issues.performance_problem import PerformanceProblem
