+++ tests/sentry/api/endpoints/test_userroles_details.py
@@ -1,6 +1,6 @@
-from sentry.models.userrole import UserRole
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.silo import control_silo_test
+from sentry.users.models.userrole import UserRole
 
 
 @control_silo_test
