+++ tests/sentry/runner/commands/test_createuser.py
@@ -1,13 +1,13 @@
 from sentry import roles
 from sentry.models.organization import Organization
 from sentry.models.organizationmember import OrganizationMember
-from sentry.models.userrole import manage_default_super_admin_role
 from sentry.receivers import create_default_projects
 from sentry.runner.commands.createuser import createuser
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import CliTestCase
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
 from sentry.users.models.user import User
+from sentry.users.models.userrole import manage_default_super_admin_role
 from sentry.users.services.user.service import user_service
